+++ api/plugins/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  auth: {
+    routes: {
+      auth: require('./auth/index'),
+    },
+    define: {}
+  },
+  ipx: require('./ipx/index'), // bu niye çalışmadı?
+}