+++ api/plugins/ipx/index.js
@@ -1,2 +1,8 @@
 const ipx = require('./middleware/ipx')
-module.exports = ipx
+module.exports = [
+  {
+    path: '/ipx',
+    method: 'get',
+    controller: ipx,
+  },
+]