+++ webpack.config.js
@@ -1,45 +1,18 @@
 const path = require('path')
+const nodeExternals = require('webpack-node-externals')
 
 // https://webpack.js.org/configuration/
 module.exports = {
   name: "server",
-  target: "node",
+  externalsPresets: { node: true }, // in order to ignore built-in modules like path, fs, etc.
+  externals: [nodeExternals()], // in order to ignore all modules in node_modules folder
   mode: 'production',
   entry: './api/index',
   output: {
     filename: 'index.js',
     path: path.resolve(__dirname, 'dist'),
     publicPath: '/',
-    // libraryTarget: 'umd',
-    // globalObject: 'this',
-  },
-  resolve: {
-    modules: [
-      "node_modules"
-    ],
-    extensions: [".ts", ".js"],
-    fallback: {
-      "tty": false,
-      "timers": false,
-      "url": false,
-      "util": false,
-      "stream": false,
-      "assert": false,
-      "os": false,
-      "crypto": false,
-      "querystring": false,
-      "http": false,
-      "path": false,
-      "fs": false,
-      "net": false,
-      "zlib": false,
-      "pg-native": false,
-      "sqlite3": false,
-      "pg-query-stream": false,
-      "oracledb": false,
-      "mysql2": false,
-      "mysql": false,
-      "tedious": false,
-    }
+    libraryTarget: 'umd',
+    globalObject: 'this',
   },
 }